pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claw-vector-svc"
version = "0.1.0"
description = "gRPC embedding microservice for the ClawDB semantic memory engine"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "sentence-transformers>=3.0",
    "onnxruntime>=1.18",
    "numpy>=1.26",
    "grpcio>=1.64",
    "grpcio-tools>=1.64",
    "protobuf>=5.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "transformers>=4.41",
    "structlog>=24.2",
    "prometheus-client>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
claw-vector-svc = "claw_vector_svc.main:main"

[tool.hatch.build.targets.wheel]
packages = ["python/claw_vector_svc", "python/proto"]

[tool.hatch.build.targets.wheel.force-include]
"python/proto/vector.proto" = "proto/vector.proto"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["python/tests"]
